requirements.txt
# Core Flask
flask>=2.3.0,<3.0
flask-cors>=4.0.0,<5.0
werkzeug>=2.3.0,<3.0

# Computer Vision & Face Recognition
opencv-python>=4.8.0
mediapipe>=0.9.0
numpy>=1.21.0
pillow>=9.0.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=1.5.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

# OTP & QR Code
pyotp>=2.9.0
qrcode>=7.4.0

# Utilities
requests>=2.28.0